% Generated by roxygen2 (4.0.0): do not edit by hand
\name{\%ha\%}
\alias{\%ha\%}
\alias{\%l*\%}
\title{Hash/Dictionary Lookup}
\usage{
terms \%ha\% key

terms \%l*\% key.match
}
\arguments{
\item{terms}{A vector of terms to undergo a lookup.}

\item{key}{The hash key to use.}
}
\description{
\code{\%ha\%} - A deprecated binary operator version of \code{hash_look}.
This will be removed in a subsequent version of \pkg{qdapTools}.  Use
\code{\%hl\%} instead.

\code{\%l*\%} - A deprecated binary operator version of \code{lookup}.  This
will be removed in a subsequent version of \pkg{qdapTools}.  Use \code{\%l\%}
instead.
}

