% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.v_outer}
\alias{print.v_outer}
\title{Prints a v_outer Object.}
\usage{
\method{print}{v_outer}(x, digits = 3, ...)
}
\arguments{
\item{x}{The v_outer object}

\item{digits}{Number of decimal places to print.}

\item{\ldots}{ignored}
}
\description{
Prints a v_outer object.
}

