% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_split.R
\name{run_split}
\alias{run_split}
\title{Split a String Into Run Chunks}
\usage{
run_split(x)
}
\arguments{
\item{x}{A string.}
}
\value{
Returns a list of vectors.
}
\description{
Splits a string into a vector or runs.
}
\examples{
run_split(c("122333444455555666666", NA, "abbcccddddeeeeeffffff"))
}
\author{
Robert Reed and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{http://stackoverflow.com/a/24319217/1000343}
}
\seealso{
Other split.functions: \code{\link{loc_split}},
  \code{\link{loc_split.character}},
  \code{\link{loc_split.data.frame}},
  \code{\link{loc_split.default}},
  \code{\link{loc_split.factor}},
  \code{\link{loc_split.list}},
  \code{\link{loc_split.matrix}},
  \code{\link{loc_split.numeric}};
  \code{\link{split_vector}}
}

