% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_end.R
\name{start_end}
\alias{start_end}
\title{Get Location of Start/End Points}
\usage{
start_end(x)
}
\arguments{
\item{x}{A vector of 1 and 0 or \code{\link[base]{logical}}.}
}
\value{
Returns a two column \code{\link[base]{data.frame}} of start and end locations for ones.
}
\description{
Get the locations of start/end places for the ones in a binary vector.
}
\examples{
set.seed(10); (x <- sample(0:1, 50, TRUE, c(.35, .65)))
start_end(x)
(y <- sample(c(TRUE, FALSE), 50, TRUE, c(.35, .65)))
start_end(y)
}
\references{
\url{https://stackoverflow.com/a/29184841/1000343}
}
\author{
Roland (\url{https://stackoverflow.com/users/1412059/roland}) and Tyler Rinker <tyler.rinker@gmail.com>.
}
\keyword{begin}
\keyword{end}
\keyword{start}
\keyword{stop}
