% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hms2sec.R
\name{hms2sec}
\alias{hms2sec}
\alias{sec2hms}
\title{Convert h:m:s To/From Seconds}
\usage{
hms2sec(x)

sec2hms(x)
}
\arguments{
\item{x}{A vector of times in h:m:s (for \code{hms2sec}) or seconds (for 
\code{sec2hms}) .}
}
\value{
\code{hms2sec} - Returns a vector of times in seconds.

\code{sec2hms} - Returns a vector of times in h:m:s format.
}
\description{
\code{hms2sec} - Converts a vector of h:m:s to seconds.

\code{sec2hms} - Converts a vector of seconds to h:m:s.
}
\examples{
hms2sec(c("02:00:03", "04:03:01"))
hms2sec(sec2hms(c(222, 1234, 55)))
sec2hms(c(256, 3456, 56565))
}
\seealso{
\code{\link[chron]{times}}
}
