% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R, R/hash.R
\name{\%l*\%}
\alias{\%l*\%}
\alias{\%ha\%}
\title{Hash/Dictionary Lookup}
\usage{
terms \%l*\% key.match

terms \%ha\% key
}
\arguments{
\item{terms}{A vector of terms to undergo a lookup.}

\item{key.match}{Takes one of the following: (1) a two column data.frame of a
match key and reassignment column, (2) a named list of vectors (Note: if
data.frame or named list supplied no key reassign needed) or (3) a single
vector match key.}

\item{key}{The hash key to use.}
}
\description{
\code{\%l*\%} - A deprecated binary operator version of \code{lookup}.  This
will be removed in a subsequent version of \pkg{qdapTools}.  Use \code{\%l\%}
instead.

\code{\%ha\%} - A deprecated binary operator version of \code{hash_look}.
This will be removed in a subsequent version of \pkg{qdapTools}.  Use
\code{\%hl\%} instead.
}
