% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_outer.R
\name{print.v_outer}
\alias{print.v_outer}
\title{Prints a v_outer Object.}
\usage{
\method{print}{v_outer}(x, digits = 3, ...)
}
\arguments{
\item{x}{The v_outer object}

\item{digits}{Number of decimal places to print.}

\item{\ldots}{ignored}
}
\description{
Prints a v_outer object.
}
